requirements.txt
arxiv==0.5.3
asgiref==3.3.0
certifi==2020.6.20
chardet==3.0.4
Django==3.1.2
django-cors-headers==3.5.0
django-filter==2.4.0
docopt==0.6.2
feedparser==6.0.2
gunicorn==20.0.4
idna==2.10
joblib==0.17.0
jsonfield==3.1.0
mosestokenizer==1.1.0
numpy==1.19.2
openfile==0.0.7
pandas==1.1.3
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
scikit-learn==0.23.2
scipy==1.5.3
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.1.0
tokenizer==2.4.0
toolwrapper==2.1.0
uctools==1.3.0
urllib3==1.25.11
whitenoise==5.2.0
xlrd==1.2.0
djangorestframework==3.12.1
django-rest-auth==0.9.5
psycopg2-binary==2.8.6



